Update AssetRenamerWindow class
- Reformat whole code
- Refactor code to use AssetRenamingOperation struct instead of three separate arrays
