+++ Editor/AssetRenamer/AssetRenamerWindow.cs.meta
@@ -1,0 +1,3 @@
+﻿fileFormatVersion: 2
+guid: b1734885d2064604b2659922d84f36c5
+timeCreated: 1630191999
