--- Editor/Custom Editors/AssetRenamerWindow.cs.meta
@@ -1,3 +1,0 @@
-﻿fileFormatVersion: 2
-guid: b1734885d2064604b2659922d84f36c5
-timeCreated: 1630191999
